Cleanup some minor type issues

+++ src/pl-index.c
@@ -350,7 +350,7 @@ static ClauseIndex
 newClauseIndexTable(int arg, hash_hints *hints)
 { GET_LD
   ClauseIndex ci = allocHeapOrHalt(sizeof(struct clause_index));
-  int m = 4;
+  unsigned int m = 4;
   size_t bytes;
 
   while(m<hints->buckets)
@@ -1214,7 +1214,7 @@ assess_remove_duplicates(hash_assessment *a, size_t clause_count)
 
 					/* assess quality */
   if ( clause_count )
-  { a->stdev   = sqrt(Q/(float)i);
+  { a->stdev   = (float)sqrt(Q/(float)i);
     a->list    = FALSE;
 
     a->speedup =            (float)(clause_count*a->size) /
@@ -1310,7 +1310,7 @@ bestHash(Word av, Definition def, hash_hints *hints)
     def->tried_index = new_bitvector(def->functor->arity);
 
 					/* Step 1: allocate assessments */
-  for(i=0; i<def->functor->arity; i++)
+  for(i=0; i<(int)def->functor->arity; i++)
   { word k;
 
     if ( !true_bit(def->tried_index, i) && (k=indexOfWord(av[i] PASS_LD)) )
@@ -1380,7 +1380,7 @@ bestHash(Word av, Definition def, hash_hints *hints)
 
   if ( best )
   { best_arg       = best->arg;
-    hints->buckets = best->size;
+    hints->buckets = (unsigned int)best->size;
     hints->speedup = best->speedup;
     hints->list    = best->list;
   }
